Update_EnhancedCharFilter
Can edit the main gallery URLs for straight and dickgirl.

Character Filter now displays the characters you choosed and the number of images you're going to copy into the filtered folder

Fixed a problem that made the character portrait to not show, that being 'cause of the change of url.