Update_NavigatorLoadPage
E621_Gallery
TagLeaderboard is not optimized and might be scratched.

E621_GalleryButton
Visual Tweaks, also added a yield for when it loads the image.

E621_MainMenu
Added the option to open the Navigation scene.

E621_Navigation
Added a "Must Have" tags.
Check the code, I'm too lazy to add details lol

E621_GlobalActions
Added the option to open the navigator,  also added a yield for when it loads the image.