Update_NewCharGal_ImprovedNavigator
E621_CharacterCreator and E621_CharacterCreatorButton:
This two new scripts open a similar Gallery System of the characters but with a lot of improvements, now you can create your own thumb for that char, and the data saved is different from the old type, not compatible with old data.

E621_Characters and E621_CharacterButton
Shold erase this

E621_Gallery
Tag Navigator now is processed by the script Data, so it's functions were removed from here.

E621_Navigation
Data downloaded for every image loaded in the navigator now is properly saved and shown to the user from the button. Improved interaction with Threads, there should be less crashes and less memory leaks, but bugs can be created.

E621_Tagger
Should rename the image if it doesn't contain the proper name, it's buggy tho.

GlobalActions
You acces the CharacterCreator from the Main Menu instead of the old Character Gallery.

ScreenshotHandler
Script that creates a screenshot taken from a camera.