Update_ArtistCreator
Data:
Created functions to save/load ArtistData (Same as CharacterData, because their properties are practically the same)

DataTypes:
Added timesSearched to characterData, might be removed.

E621_CharacterCreator:
The script has been adapted to be used as an ArtistCreator by an enum.

E621_NavigationButton:
Added an special button that shows the Characters/Artists (also specificTags but was commented) of the image shown in the navigator, it's function is to show which character or artist is already registered in the database.

GlobalActions:
Added "artists" scene compatibility to the OpenSceneAsync()

Some prefabs were modified and resources added.