Update_Arist/CharacterCreatorTweaks
ButtonCharacterCreator.prefab
Attachments are smaller, edit button moved down and there's a new stats text.

Data
Tag Selector now has functions for Artists Tags.

CharacterCreator
A new queue was created for the display of the info for the Character/Artist buttons. It's queued for not oversaturating the Threaded functions.

CharacterCreatorButton
Thread for getting the stats of the Character/Artist (Images, Filtered and Ratio) was added.